got a better idea
