--- Assets/Scripts/Dialogue_Scripts/NewBulletinBoard.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 54f9fdf4acce0c64397d318dc36a41f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
